Update config.toml
I had commented out the image and thumb links for the portfolio gallery; however, broken links (gray images) still appeared in the mobile view on my smartphone. I have now deleted the image and thumb code in an effort to make the portfolio site work on my smartphone. (Not clear if this solves the issue but I have manually removed the href lines from index.html for the actual site as a workaround.